requirements.txt
numpy==2.2.6
orjson==3.11.5
sortedcontainers==2.4.0
aiofiles==25.1.0
picows==1.10.2
python-dotenv==1.1.0
paradex_py==0.5.4
lighter-sdk==1.0.2
hyperliquid-python-sdk==0.23.0
python_binance==1.0.36